+++ src/trabajadores/trabajadores.controller.ts
@@ -15,10 +15,14 @@ import {
 } from "./trabajadores.dto";
 
 import { RoleGuard } from "../guards/role.guard";
+import { LoggerService } from "src/logger/logger.service";
 
 @Controller("trabajadores")
 export class TrabajadoresController {
-  constructor(private readonly trabajadorInstance: TrabajadorService) {}
+  constructor(
+    private readonly trabajadorInstance: TrabajadorService,
+    private readonly loggerService: LoggerService,
+  ) {}
 
   @UseGuards(AuthGuard)
   @Get()
@@ -155,8 +159,21 @@ export class TrabajadoresController {
     @Body("modificado")
     modificado: TrabajadorFormRequest,
     // @User() firebaseUser: UserRecord,
+    @User() user: UserRecord,
   ) {
     try {
+      const usuarioCompleto =
+        await this.trabajadorInstance.getTrabajadorByAppId(user.uid);
+      const nombreUsuario = usuarioCompleto?.nombreApellidos || user.email;
+      // Registro de auditoría
+      await this.loggerService.create({
+        action: "Actualizar Trabajador",
+        name: nombreUsuario,
+        extraData: {
+          originalData: original,
+          modifiedData: modificado,
+        },
+      });
       return {
         ok: true,
         data: await this.trabajadorInstance.guardarCambiosForm(
@@ -242,7 +259,30 @@ export class TrabajadoresController {
   @Roles("Super_Admin", "RRHH_ADMIN")
   @UseGuards(AuthGuard, RoleGuard)
   @Post("eliminar")
-  async eliminarTrabajador(@Body() req: DeleteTrabajadorDto) {
+  async eliminarTrabajador(
+    @Body() req: DeleteTrabajadorDto,
+    @User() user: UserRecord,
+  ) {
+    const trabajadorDelete = await this.trabajadorInstance.getTrabajadorBySqlId(
+      req.id,
+    );
+    if (!trabajadorDelete) {
+      throw new Error("Trabajador no encontrada");
+    }
+
+    // 3. Obtener el nombre del usuario autenticado
+    const usuarioCompleto = await this.trabajadorInstance.getTrabajadorByAppId(
+      user.uid,
+    );
+    const nombreUsuario = usuarioCompleto?.nombreApellidos || user.email;
+
+    // 4. Registrar la auditoría con la información del trabajador eliminado
+    await this.loggerService.create({
+      action: "Eliminar Trabajador",
+      name: nombreUsuario,
+      extraData: { trabajadorData: trabajadorDelete },
+    });
+
     return await this.trabajadorInstance.eliminarTrabajador(req.id);
   }
